feat: enable jwt authentication

+++ src/auth/auth.controller.ts
@@ -7,6 +7,7 @@ import {
   UseGuards,
   Request,
   Get,
+  ForbiddenException,
 } from '@nestjs/common';
 import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
 import { AuthService } from './auth.service';
@@ -18,6 +19,7 @@ import { AuthGuard } from '@nestjs/passport';
 import { WalletAuthStrategy } from './wallet-auth.strategy';
 import { PasswordAuthStrategy } from './password-auth.strategy';
 import { JwtAuthGuard } from './jwt-guard.guard';
+import { UserDocument } from '../user/entities/user.entity';
 
 @ApiTags('auth')
 @Controller('api/auth')
@@ -56,13 +58,29 @@ export class AuthController {
   }
 
   @UseGuards(JwtAuthGuard)
-  @Post('/connect-wallet')
-  connectWallet(@Body() createAuthDto: CreateAuthDto) {
+  @Post('connect-wallet')
+  connectWallet(@Body() createAuthDto: CreateAuthDto, @Request() req: Request) {
+    const { user } = req as unknown as { user: UserDocument };
+
+    if (user.id.toString() !== createAuthDto.userId) {
+      throw new ForbiddenException();
+    }
+
     return this.authService.createAuthEntity(createAuthDto);
   }
 
-  @Delete(':auth_id')
-  deleteEntity(@Param('auth_id') id: string) {
+  @Delete(':user_id/:auth_id')
+  @UseGuards(JwtAuthGuard)
+  deleteEntity(
+    @Param('auth_id') id: string,
+    @Param('user_id') user_id: string,
+    @Request() req: Request,
+  ) {
+    const { user } = req as unknown as { user: UserDocument };
+
+    if (user.id.toString() !== user_id) {
+      throw new ForbiddenException();
+    }
     return this.authService.deleteAuthEntity(id);
   }
 }
